release pipeline python client (#83)

+++ clients/python/setup.py
@@ -7,6 +7,7 @@ import urllib.request
 import tarfile
 import zipfile
 import platform
+import json
 
 from setuptools import setup
 
@@ -28,27 +29,35 @@ DIR_PATH = Path(__file__).parent
 ROOT_PATH = DIR_PATH.joinpath("..").joinpath("..")
 # path to where the c-wrapper code for the surrealML core is
 C_PATH = ROOT_PATH.joinpath("modules").joinpath("c-wrapper")
-# path to where the binary is after building the c-wrapper
+# path to where the binary is if it's in c-wrapper
 BINARY_PATH = C_PATH.joinpath("target").joinpath("release").joinpath(get_c_lib_name())
+# path to where the binary is if it's in the root
+ROOT_BINARY_PATH = ROOT_PATH.joinpath("target").joinpath("release").joinpath(get_c_lib_name())
 # path to where the c-wrapper is stored if held in the surrealML python package
 BINARY_DIST = DIR_PATH.joinpath("surrealml").joinpath(get_c_lib_name())
-# path to where the onnxruntime is stored if held in the surrealML python package
-ONNXRUNTIME_DIR = DIR_PATH.joinpath("surrealml").joinpath("onnxruntime")
 
-# get information about the system
-ONNX_VERSION = "1.20.0"
-PYTHON_PACKAGE_VERSION = "0.1.0"
+# get the python package version
+CONFIG_JSON_PATH = Path(__file__).parent.joinpath("config.json")
+
+def read_version():
+    try:
+        with open(CONFIG_JSON_PATH, "r") as json_data:
+            config = json.load(json_data)
+            return config["version"]
+    except Exception as e:
+        print(f"Error loading version from '{CONFIG_JSON_PATH}': {e}", file=sys.stderr)
+        sys.exit(1)
+
+# Allow CI to override target platform for downloads:
+OVERRIDE_OS = os.environ.get("TARGET_OS")
+OVERRIDE_ARCH = os.environ.get("TARGET_ARCH")
+
+# save vars about the system
+PYTHON_PACKAGE_VERSION = read_version()
 DYNAMIC_LIB_VERSION = "0.1.0"
-OS_NAME = sys.platform
-ARCH = platform.machine().lower()
+OS_NAME = OVERRIDE_OS or sys.platform  
+ARCH = (OVERRIDE_ARCH or platform.machine()).lower()
 SYSTEM = platform.system().lower()  # 'linux', 'darwin' (macOS), 'windows'
-
-if OS_NAME.startswith("linux"):
-    ONNX_LIB_NAME = "libonnxruntime.so"
-elif OS_NAME == "darwin":
-    ONNX_LIB_NAME = "libonnxruntime.dylib"
-elif OS_NAME == "win32":
-    ONNX_LIB_NAME = "libonnxruntime.dll"
 
 ROOT_DEP_DIR = os.path.expanduser("~/surrealml_deps")
 
@@ -56,40 +65,12 @@ ROOT_DEP_DIR = os.path.expanduser("~/surrealml_deps")
 os.makedirs(ROOT_DEP_DIR, exist_ok=True)
 
 # Paths for versioned libraries
-ONNX_LIB_DIR = os.path.join(ROOT_DEP_DIR, "onnxruntime", ONNX_VERSION)
 DYNAMIC_LIB_DIR = os.path.join(ROOT_DEP_DIR, "core_ml_lib", DYNAMIC_LIB_VERSION)
-ONNX_DOWNLOAD_CACHE = os.path.join(ONNX_LIB_DIR, "download_cache.tgz")
 DYNAMIC_LIB_DIST = DIR_PATH.joinpath(DYNAMIC_LIB_DIR).joinpath(get_c_lib_name())
 DYNAMIC_LIB_DOWNLOAD_CACHE = os.path.join(DYNAMIC_LIB_DIR, "download_cache.tgz")
 
 # create the directories if they don't exist
-os.makedirs(ONNX_LIB_DIR, exist_ok=True)
 os.makedirs(DYNAMIC_LIB_DIR, exist_ok=True)
-
-# ===================================== download the onnxruntime into the dep directory =========================================================
-
-def get_onnxruntime_url():
-    base_url = f"https://github.com/microsoft/onnxruntime/releases/download/v{ONNX_VERSION}/"
-
-    if OS_NAME.startswith("linux"):
-        if ARCH == "x86_64":
-            return f"{base_url}onnxruntime-linux-x64-{ONNX_VERSION}.tgz", f"onnxruntime-linux-x64-{ONNX_VERSION}"
-        elif ARCH in ("arm64", "aarch64"):
-            return f"{base_url}onnxruntime-linux-aarch64-{ONNX_VERSION}.tgz", f"onnxruntime-linux-aarch64-{ONNX_VERSION}"
-    
-    elif OS_NAME == "darwin":
-        if ARCH == "x86_64":
-            return f"{base_url}onnxruntime-osx-x86_64-{ONNX_VERSION}.tgz", f"onnxruntime-osx-x86_64-{ONNX_VERSION}"
-        elif ARCH == "arm64":
-            return f"{base_url}onnxruntime-osx-arm64-{ONNX_VERSION}.tgz", f"onnxruntime-osx-arm64-{ONNX_VERSION}"
-
-    elif OS_NAME == "win32":
-        if ARCH == "x86_64":
-            return f"{base_url}onnxruntime-win-x64-{ONNX_VERSION}.zip", f"onnxruntime-win-x64-{ONNX_VERSION}"
-        elif ARCH == "arm64":
-            return f"{base_url}onnxruntime-win-arm64-{ONNX_VERSION}.zip", f"onnxruntime-win-arm64-{ONNX_VERSION}"
-    
-    raise Exception(f"Unsupported platform or architecture: {OS_NAME}")
 
 
 def get_lib_url():
@@ -114,25 +95,6 @@ def get_lib_url():
     raise Exception(f"Unsupported platform or architecture: {OS_NAME}")
 
 
-def download_and_extract_onnxruntime():
-    url, extracted_dir = get_onnxruntime_url()
-
-    if not os.path.exists(ONNX_DOWNLOAD_CACHE):
-        print(f"Downloading ONNX Runtime from {url}")
-        urllib.request.urlretrieve(url, ONNX_DOWNLOAD_CACHE)
-    else:
-        print(f"the {ONNX_DOWNLOAD_CACHE} alread exists so not downloading onnxruntime")
-
-    # Extract based on file type
-    if ONNX_DOWNLOAD_CACHE.endswith(".tgz"):
-        with tarfile.open(ONNX_DOWNLOAD_CACHE, "r:gz") as tar:
-            tar.extractall(path=ONNX_LIB_DIR)
-    elif ONNX_DOWNLOAD_CACHE.endswith(".zip"):
-        with zipfile.ZipFile(ONNX_DOWNLOAD_CACHE, "r") as zip_ref:
-            zip_ref.extractall(ONNX_LIB_DIR)
-
-    return extracted_dir
-
 def download_and_extract_c_lib():
     url, extracted_dir = get_lib_url()
     if not os.path.exists(DYNAMIC_LIB_DOWNLOAD_CACHE):
@@ -147,20 +109,10 @@ def download_and_extract_c_lib():
     elif DYNAMIC_LIB_DOWNLOAD_CACHE.endswith(".zip"):
         with zipfile.ZipFile(DYNAMIC_LIB_DOWNLOAD_CACHE, "r") as zip_file:
             zip_file.extractall(path=DYNAMIC_LIB_DIR)
+
     return extracted_dir
 
 
-
-onnx_lib_dist = Path(ONNX_LIB_DIR).joinpath(ONNX_LIB_NAME)
-
-# downloads and unpacks the onnx lib if the onnx lib is not present
-if os.path.exists(onnx_lib_dist) is False:
-    print("downloading the path")
-    onnxruntime_path = download_and_extract_onnxruntime()
-    onnx_lib_path = Path(ONNX_LIB_DIR).joinpath(onnxruntime_path).joinpath("lib").joinpath(ONNX_LIB_NAME)
-    shutil.copyfile(onnx_lib_path, onnx_lib_dist)
-    shutil.rmtree(Path(ONNX_LIB_DIR).joinpath(onnxruntime_path))
-    os.remove(ONNX_DOWNLOAD_CACHE)
 
 # ===================================== Build the rust binary into a dynamic C lib =======================================
 
@@ -170,7 +122,13 @@ BUILD_FLAG = False
 if DYNAMIC_LIB_DIST.exists() is False and os.environ.get("LOCAL_BUILD") == "TRUE":
     print("building core ML lib locally")
     subprocess.Popen("cargo build --release", cwd=str(C_PATH), shell=True).wait()
-    shutil.copyfile(BINARY_PATH, DYNAMIC_LIB_DIST)
+    ARCH = (OVERRIDE_ARCH or platform.machine()).lower()
+    
+    if BINARY_PATH.exists():
+        shutil.copyfile(BINARY_PATH, DYNAMIC_LIB_DIST)
+    elif ROOT_BINARY_PATH.exists():
+        shutil.copyfile(ROOT_BINARY_PATH, DYNAMIC_LIB_DIST)
+
     BUILD_FLAG = True
 
 else:
@@ -183,6 +141,16 @@ else:
         # DYNAMIC_LIB_DIST
         # DYNAMIC_LIB_DOWNLOAD_CACHE
 
+        # build path to the freshly-extracted library
+        downloaded_file = Path(DYNAMIC_LIB_DIR) / get_c_lib_name()
+        if not downloaded_file.exists():
+            raise Exception(f"Expected shared lib at {downloaded_file}, but none was found")
+
+        # copy it into the package so package_data will include it
+        BINARY_DIST.parent.mkdir(parents=True, exist_ok=True)
+        shutil.copyfile(downloaded_file, BINARY_DIST)
+
+
 # ===================================== run the setup for the python package =============================================
 
 setup(
@@ -191,7 +159,7 @@ setup(
     description="A machine learning package for interfacing with various frameworks.",
     author="Maxwell Flitton",
     author_email="maxwellflitton@gmail.com",
-    url="https://github.com/surrealdb/surrealml",
+    url="https://github.com/surrealdb/surrealml"
     license="MIT",
     classifiers=[
         "Programming Language :: Python :: 3",
@@ -203,6 +171,9 @@ setup(
         "numpy==1.26.3",
     ],
     extras_require={
+        "dev": [
+            "twine",
+        ],
         "sklearn": [
             "skl2onnx==1.16.0",
             "scikit-learn==1.4.0",
@@ -215,6 +186,12 @@ setup(
             "tensorflow==2.16.1",
         ],
     },
+    include_package_data=True,   
+    zip_safe=False,
+    package_data={"surrealml": ["*.so", "*.dylib", "*.dll"]},   
+    data_files=[
+        ("", ["config.json"]),      
+    ],
     packages=[
         "surrealml",
         "surrealml.engine",
